pyproject.toml
[project]
name = "athleticspose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.0.1",
    "torchvision==0.15.2",
    "matplotlib>=3.10.0",
    "natsort>=8.4.0",
    "opencv-python>=4.11.0.86",
    "plotly>=6.0.0",
    "seaborn>=0.13.2",
    "numpy<2",
    "lightning==2.2",
    "wandb>=0.19.7",
    "timm>=1.0.15",
    "chumpy==0.66",
    "mmengine>=0.10.5",
    "mmcv",
    "mmdet>=3.3.0",
    "mmpose>=1.3.1",
    "mmpretrain>=1.2.0",
    "hydra-core>=1.3.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.4",
]
exp = [
    "jupyter>=1.1.1",
    "openpyxl>=3.1.5",
    "scikit-learn>=1.7.0",
    "statsmodels>=0.14.5",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["athleticspose"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'"}
torchvision = { index = "pytorch", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'"}
mmcv = { url = "https://download.openmmlab.com/mmcv/dist/cu118/torch2.0.0/mmcv-2.1.0-cp310-cp310-manylinux1_x86_64.whl" }

[tool.ruff.lint]
unfixable = ["F401"]
